pyproject.toml
[tool.poetry]
name = "ctfd-sdk"
version = "0.1.0"
description = "Python SDK for CTFd REST API"
authors = ["Simon Plhak <plhak.s@gmail.com>"]
readme = "README.md"
packages = [{include = "ctfd_sdk"}]

[tool.poetry.dependencies]
python = "^3.9"
httpx = "^0.27.0"
python-dotenv = "^1.0.1"


[tool.poetry.group.dev.dependencies]
black = "^24.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120
skip-string-normalization = true
extend-exclude = '.*/env/'
